+++ CHANGELOG.md
@@ -8,9 +8,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### v3.9.x - [Unreleased]
 
+### v3.9.8.5 - 2019-08-29
+
 #### Fixed
 
 - Fix Netdata install on Raspbian 9/10
+- `wo stack remove/purge` confirmation
+- Nginx error after removing a SSL certificate used to secure WordOps backend
 
 ### v3.9.8.4 - 2019-08-28